Modified configuration loading mechanism.
Renamed *ConfigLoader to *PropertiesLoader.
Introduced PropertyLoader and DefaultPropertyLoader
Introduced ProxyConfig which contains just the proxy config settings
Renamed package to org.concordion.cubano.config
Made Config and WebDriverConfig final
Removed inheritance from Config and WebDriverConfig
Introduced singletons for DefaultPropertiesLoader, Config and WebDriverConfig to ensure that properties files are only read once, and make it easy to access Config and WebDriverConfig
Removed HttpEasyConfig, since this just uses ProxyConfig
Modified WebDriverConfig to mirror Config and delegate to PropertyLoader

Any specific AppConfig, it will no longer be able to inherit from Config or WebDriverConfig and will need to be modified to delegate to PropertyLoader. See WebDriverConfig for example code.
