+++ cubano-config/src/main/java/org/concordion/cubano/config/CaselessProperties.java
@@ -1,4 +1,4 @@
-package org.concordion.cubano.utils;
+package org.concordion.cubano.config;
 
 import java.util.Map;
 import java.util.Properties;
