+++ cubano-config/src/main/java/org/concordion/cubano/config/DefaultPropertiesLoader.java
@@ -1,0 +1,73 @@
+package org.concordion.cubano.config;
+
+import java.io.File;
+import java.io.StringReader;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.Properties;
+
+/**
+ * Loads configuration from config.properties and user.properties located in projects root folder.
+ * 
+ * It also ensures that any single slash character '\' does not require a double slash. This means that
+ * using \ to wrap long property values across multiple lines won't work.
+ */
+public class DefaultPropertiesLoader implements PropertiesLoader {
+    private static final String CONFIG_FILE = "config.properties";
+    private static final String USER_CONFIG_FILE = "user.properties";
+
+    private final Properties properties;
+    private final Properties userProperties;
+
+    private static class DPHolder {
+        static final DefaultPropertiesLoader INSTANCE = new DefaultPropertiesLoader();
+    }
+
+    public static DefaultPropertiesLoader getInstance() {
+        return DPHolder.INSTANCE;
+    }
+
+    /** Ensure properties have been loaded before any property is used. */
+    private DefaultPropertiesLoader() {
+        synchronized (DefaultPropertiesLoader.class) {
+            properties = loadFile(CONFIG_FILE);
+
+            if (new File(USER_CONFIG_FILE).exists()) {
+                userProperties = loadFile(USER_CONFIG_FILE);
+            } else {
+            	userProperties = null;
+            }
+        }
+    }
+
+    /**
+     * Read properties from file, will ignoring the case of properties.
+     *
+     * @param filename Name of file to read, expected that it will be located in the projects root folder
+     * @return {@link CaselessProperties}
+     */
+    private Properties loadFile(final String filename) {
+        Properties prop = new CaselessProperties();
+
+        try {
+            String content = new String(Files.readAllBytes(Paths.get(filename)));
+
+            // By default property files treat \ as an escape character
+            prop.load(new StringReader(content.replace("\\", "\\\\")));
+        } catch (Exception e) {
+            throw new RuntimeException("Unable to read properties file.", e);
+        }
+
+        return prop;
+    }
+
+    @Override
+    public Properties getProperties() {
+        return properties;
+    }
+
+    @Override
+    public Properties getUserProperties() {
+        return userProperties;
+    }
+}
