+++ cubano-config/src/main/java/org/concordion/cubano/config/DefaultPropertyLoader.java
@@ -1,0 +1,192 @@
+package org.concordion.cubano.config;
+
+import java.util.Enumeration;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Properties;
+
+/**
+ * A PropertyLoader which loads a property, firstly checking the userProperties and then the properties.
+ * Within each set of properties, it firstly checks for a key matching <i>environment.key</i> and then for <i>key</i>
+ * where <i>environment</i> is set using setEnvironment and <i>key</i> is the key parameter.
+ */
+public class DefaultPropertyLoader implements PropertyLoader {
+    private final Properties properties;
+    private final Properties userProperties;
+    private String environment;
+
+	/**
+	 * Configure the optional userProperties and mandatory properties to be loaded.
+	 */
+	public DefaultPropertyLoader(Properties properties, Properties userProperties) {
+        this.properties = properties;
+        this.userProperties = userProperties;
+    }
+
+	/**
+	 * Sets the environment that this property loader will use as a prefix for properties.
+	 * @param environment the environment to use as a prefix when looking for a property
+	 */
+	public void setEnvironment(String environment) {
+		this.environment = environment;
+	}
+
+    /**
+	 * Get the property for the current environment, if that is not found it will look for "{@literal <key>}".
+	 *
+	 * @param key Id of the property to look up
+	 * @return Property value if found, throws exception if not found
+	 */
+	@Override
+	public String getProperty(String key) {
+		String value = retrieveProperty(key);
+
+		if (value.isEmpty()) {
+			throw new IllegalArgumentException(String.format("Unable to find property %s", key));
+		}
+
+		return value;
+	}
+
+    /**
+	 * Get the property for the current environment, if that is not found it will look for "{@literal <key>}".
+	 *
+	 * @param key Id of the property to look up
+	 * @param defaultValue value to use if property is not found
+	 * @return Property value if found, defaultValue if not found
+	 */
+	@Override
+	public String getProperty(String key, String defaultValue) {
+		String value = retrieveProperty(key);
+
+		if (value.isEmpty()) {
+			value = defaultValue == null ? "" : defaultValue;
+		}
+
+		return value;
+	}
+
+    /**
+	 * Get the property for the current environment as a boolean value, if that is not found it will look for "{@literal <key>}".
+	 *
+	 * @param key Id of the property to look up
+	 * @param defaultValue value to use if property is not found
+	 * @return Property value if found, defaultValue if not found
+	 */
+	@Override
+	public boolean getPropertyAsBoolean(String key, String defaultValue) {
+		String value = retrieveProperty(key);
+
+		if (value.isEmpty()) {
+			value = defaultValue == null ? "false" : defaultValue;
+		}
+
+		return Boolean.valueOf(value);
+	}
+
+    /**
+	 * Get the property for the current environment as a numeric value, if that is not found it will look for "{@literal <key>}".
+	 *
+	 * @param key Id of the property to look up
+	 * @param defaultValue value to use if property is not found
+	 * @return Property value if found, defaultValue if not found
+	 */
+	@Override
+	public int getPropertyAsInteger(String key, String defaultValue) {
+		String value = retrieveProperty(key);
+
+		if (value.isEmpty()) {
+			value = defaultValue == null ? "0" : defaultValue;
+		}
+
+		return Integer.valueOf(value);
+	}
+
+    /**
+	 * Returns a map of key value pairs of properties starting with a prefix.
+	 *
+	 * @param keyPrefix Search string
+	 * @return Map
+	 */
+	@Override
+	public Map<String, String> getPropertiesStartingWith(String keyPrefix) {
+		return getPropertiesStartingWith(keyPrefix, false);
+	}
+
+    /**
+	 * Returns a map of key value pairs of properties starting with a prefix.
+	 *
+	 * @param keyPrefix Search string
+	 * @param trimPrefix Remove prefix from key in returned set
+	 * @return Map
+	 */
+	@Override
+	public Map<String, String> getPropertiesStartingWith(String keyPrefix, boolean trimPrefix) {
+		Map<String, String> result = new HashMap<>();
+
+		searchPropertiesFrom(properties, keyPrefix, trimPrefix, result);
+		searchPropertiesFrom(userProperties, keyPrefix, trimPrefix, result);
+
+		return result;
+	}
+
+    private void searchPropertiesFrom(Properties properties, String keyPrefix, boolean trimPrefix, Map<String, String> result) {
+		if (properties == null) {
+			return;
+		}
+
+		@SuppressWarnings("unchecked")
+        Enumeration<String> en = (Enumeration<String>) properties.propertyNames();
+		while (en.hasMoreElements()) {
+			String propName = en.nextElement();
+			String propValue = properties.getProperty(propName);
+
+			if (propName.startsWith(keyPrefix)) {
+				if (trimPrefix) {
+					propName = propName.substring(keyPrefix.length());
+				}
+
+				result.put(propName, propValue);
+			}
+		}
+	}
+
+    private String retrieveProperty(String key) {
+		String value = null;
+
+		// prefix = System.getProperty("user.name").toLowerCase();
+		if (userProperties != null) {
+			value = retrievePropertyFrom(userProperties, key);
+		}
+
+		if (value == null) {
+			value = retrievePropertyFrom(properties, key);
+		}
+
+		if (value == null) {
+			value = "";
+		}
+
+		return value;
+	}
+
+    private String retrievePropertyFrom(Properties properties, String key) {
+		String value = null;
+
+		// Attempt to get setting for environment
+		if (environment != null && !environment.isEmpty()) {
+			value = properties.getProperty(environment + "." + key);
+		}
+
+		// Attempt to get default setting
+		if (value == null) {
+			value = properties.getProperty(key);
+		}
+
+		if (value != null) {
+			value = value.trim();
+		}
+
+		return value;
+	}
+}
