+++ cubano-core/src/main/java/org/concordion/cubano/config/TimeoutException.java
@@ -1,4 +1,4 @@
-package org.concordion.cubano.utils;
+package org.concordion.cubano.config;
 
 /**
  * Timeout Exception.
