+++ cubano-core/src/main/java/org/concordion/cubano/config/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Common utilties that most automation projects require.
  */
-package org.concordion.cubano.utils;
+package org.concordion.cubano.config;
