+++ cubano-core/src/test/java/org/concordion/cubano/config/ISODateTimeFormatTest.java
@@ -1,4 +1,4 @@
-package org.concordion.cubano.utils;
+package org.concordion.cubano.config;
 
 import org.junit.Test;
