+++ cubano-webdriver-manager/config.properties
@@ -2,14 +2,14 @@
 
 environment = dev
 
-#webdriver.browserprovider = ChromeBrowserProvider
+webdriver.browserprovider = ChromeBrowserProvider
 #webdriver.browserprovider = EdgeBrowserProvider
 #webdriver.browserprovider = FirefoxBrowserProvider
 #webdriver.browserprovider = InternetExplorerBrowserProvider
-webdriver.browserprovider = OperaBrowserProvider
+#webdriver.browserprovider = OperaBrowserProvider
 #webdriver.browserprovider = SafariBrowserProvider
 
-opera.exe = C:\Program Files\Opera\48.0.2685.39\opera.exe
+#opera.exe = C:\Program Files\Opera\48.0.2685.39\opera.exe
 #ie.capability.initialBrowserUrl = http://concordion.org
 #ie.capability.requireWindowFocus = true
