+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,6 @@
+#Mon Oct 23 10:57:04 NZDT 2017
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.2.1-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.2.1-all.zip
