+++ structure101-settings.javax.hsp
@@ -16,49 +16,29 @@
     <classpathentry kind="lib" path="SEPARATOR" />
     <classpathentry kind="lib" path="classes/production/cubano-core" base="cubano-core" />
     <classpathentry kind="lib" path="SEPARATOR" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
+    <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
+    <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
     <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
     <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
-    <classpathentry kind="lib" path="cubano-httpeasy/build/classes/java/main" base="cubano-httpeasy_main" />
-    <classpathentry kind="lib" path="SEPARATOR" />
-    <classpathentry kind="lib" path="cubano-httpeasy/build/classes/java/main" base="cubano-httpeasy_main" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
-    <classpathentry kind="lib" path="SEPARATOR" />
-    <classpathentry kind="lib" path="cubano-webdriver/build/classes/java/main" base="cubano-webdriver_main" />
     <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
     <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
-    <classpathentry kind="lib" path="cubano-webdriver/build/classes/java/main" base="cubano-webdriver_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
-    <classpathentry kind="lib" path="cubano-webdriver-manager/build/classes/java/main" base="cubano-webdriver-manager_main" />
     <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
-    <classpathentry kind="lib" path="cubano-webdriver/build/classes/java/main" base="cubano-webdriver_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
-    <classpathentry kind="lib" path="cubano-webdriver-manager/build/classes/java/main" base="cubano-webdriver-manager_main" />
     <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
-    <classpathentry kind="lib" path="cubano-webdriver/build/classes/java/main" base="cubano-webdriver_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
-    <classpathentry kind="lib" path="cubano-core/build/classes/java/main" base="cubano-core_main" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
     <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
-    <classpathentry kind="lib" path="cubano-webdriver/build/classes/java/main" base="cubano-webdriver_main" />
-    <classpathentry kind="lib" path="cubano-httpeasy/build/classes/java/main" base="cubano-httpeasy_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
-    <classpathentry kind="lib" path="cubano-http/build/classes/java/main" base="cubano-http_main" />
     <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
-    <classpathentry kind="lib" path="cubano-core/build/classes/java/main" base="cubano-core_main" />
-    <classpathentry kind="lib" path="cubano-webdriver/build/classes/java/main" base="cubano-webdriver_main" />
-    <classpathentry kind="lib" path="cubano-httpeasy/build/classes/java/main" base="cubano-httpeasy_main" />
+    <classpathentry kind="lib" path="SEPARATOR" />
+    <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
+    <classpathentry kind="lib" path="SEPARATOR" />
+    <classpathentry kind="lib" path="cubano-config/build/classes/java/main" base="cubano-config_main" />
     <classpathentry kind="lib" path="SEPARATOR" />
   </classpath>
   <pom-root-files />
